#231 Refactor booking validation to BookingService
Moved booking request validation logic from command handlers to a new BookingService class for better separation of concerns and reusability. Updated dependency injection and handler constructors to use the new service. Added missing Json serialization import in CreateRequestBookingDto.
